pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-archive"
version = "0.1.0"
description = "Capture, organize and sync ChatGPT and Gemini conversations"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.44",
    "pydantic>=2.7",
    "typer>=0.12",
    "beautifulsoup4>=4.12",
    "markdownify>=0.12",
    "tenacity>=8.3",
    "orjson>=3.10",
    "python-dotenv>=1.0",
    "sentence-transformers>=3.0",
    "scikit-learn>=1.4",
    "rapidfuzz>=3.8",
    "google-api-python-client>=2.130",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "rich>=13.7",
    "pyyaml>=6.0",
    "keybert>=0.8",
    "hdbscan>=0.8",
    "yake>=0.4",
    "sqlmodel>=0.0.18",
    "markdown>=3.6",
]

[project.optional-dependencies]
dev = ["pytest>=8.2", "pytest-asyncio>=0.23"]

[project.scripts]
ai-archive = "ai_archive.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_archive"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
